chore(test): extract apple legacy feature test
