+++ tests/Feature/HandleAppStoreNotificationFeatureTest.php
@@ -6,7 +6,6 @@ namespace Imdhemy\Purchases\Tests\Feature;
 
 use Illuminate\Support\Facades\Event;
 use Imdhemy\AppStore\Jws\JwsVerifier;
-use Imdhemy\Purchases\Events\AppStore\DidChangeRenewalStatus;
 use Imdhemy\Purchases\Events\AppStore\Subscribed;
 use Imdhemy\Purchases\Tests\TestCase;
 
@@ -17,22 +16,6 @@ final class HandleAppStoreNotificationFeatureTest extends TestCase
         parent::setUp();
 
         $this->app->bind(JwsVerifier::class, \Imdhemy\Purchases\Tests\Doubles\JwsVerifier::class);
-    }
-
-    /** @test */
-    public function handle_app_store_server_notifications(): void
-    {
-        $this->withoutExceptionHandling();
-        Event::fake();
-        $data = json_decode(
-            file_get_contents($this->fixturesDir('appstore-server-notification.json')),
-            true,
-            512,
-            JSON_PARTIAL_OUTPUT_ON_ERROR
-        );
-        $this->post('/liap/notifications?provider=app-store', $data)->assertStatus(200);
-
-        Event::assertDispatched(DidChangeRenewalStatus::class);
     }
 
     /** @test */
